try manual sharding
