+++ .github/workflows/main.yml
@@ -57,7 +57,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        shard: [1, 2, 3, 4]
+        shard: [1, 2]
     steps:
       - uses: actions/checkout@v4
 
@@ -72,6 +72,7 @@ jobs:
         run: yarn
 
       - name: Test
+
         run: yarn test --shard=${{ matrix.shard }}/${{ strategy.job-total }} --coverage
 
       - run: mv coverage/coverage-final.json coverage/${{matrix.shard}}.json
