+++ src/test/checker.e2e.0.test.ts
@@ -1,0 +1,3 @@
+import {checkerTest} from './checker-logic.js';
+
+checkerTest(0, 2);
