+++ src/test/checker.e2e.1.test.ts
@@ -1,0 +1,3 @@
+import {checkerTest} from './checker-logic.js';
+
+checkerTest(1, 2);
